Recognize vegetarian and vegan alternatives
These seem to be offered sometimes, such that it's only noted in the name of the
meal that they exist (clearly with a non-vegan resp. vegetarian option it would
be wrong to declare the whole meal as such).

There's also some cleanup and refactoring here that I can't be bothered to put
into a separate commit.