Add opening-times command
This is very rough right now, but works well enough already that I want
to include it.

Still missing are respecting the users' wishes in terms of line wraps et
al, as well as queries of all sorts: for a specific canteen, as well as
what is open at a specific date. This will probably force me to actually
parse the HTML site, which is also a plus. That's also why there's not
man page update for now.

Another thing is that I'm not sure whether the split into a new Mode
type makes sense at all. Some settings, like wrapping, should surely
carry over to all settings, and not just to these, right?

All that is to say, it's the weekend and I don't want to think about all
that right now :)